Removed redundant interface
